requirements.txt
Pillow==11.3.0
python-dotenv==1.0.1
